+++ BasicModels/Board.cs
@@ -25,8 +25,12 @@ namespace battleship
         private static readonly string[] COLUMN_HEADERS = { "1", "2", "3", "4", "5", "6", "7", "8", "9", "10" };
         public static int ROW_OFFSET = 12;
         public static int COLUMN_OFFSET = 50;
+        private static string EMPTY_SPACE = $"|{Display.WriteUnderline(" ")}";
+        private static string EMPTY_RIGHT_EDGE_SPACE = $"|{Display.WriteUnderline(" ")}|";
+        private static string BOAT_SPACE = $"|{Display.WriteUnderline("O")}";
+        private static string BOAT_SPACE_RIGHT_EDGE = $"|{Display.WriteUnderline("O")}|";
 
-        public Boat[] Boats { get; set; } 
+        public Boat[] Boats { get; set; }
         public string[,] GameBoard { get; set; }
 
         public Board()
@@ -64,7 +68,7 @@ namespace battleship
         {
             for (int row = 0; row < BOARD_DIMENSION - 1; row++)
             {
-               GameBoard[row + 1, 0] = $" {ROW_HEADERS[row]} ";
+                GameBoard[row + 1, 0] = $" {ROW_HEADERS[row]} ";
             }
         }
 
@@ -82,11 +86,148 @@ namespace battleship
             {
                 for (int column = 1; column < BOARD_DIMENSION; column++)
                 {
-                    if (column == BOARD_DIMENSION - 1) GameBoard[row, column] = $"|{Display.WriteUnderline(" ")}|";
-                    else GameBoard[row, column] = $"|{Display.WriteUnderline(" ")}";
+                    if (column == BOARD_DIMENSION - 1) GameBoard[row, column] = EMPTY_RIGHT_EDGE_SPACE;
+                    else GameBoard[row, column] = EMPTY_SPACE;
                 }
             }
         }
 
+        public bool AddBoatToBoard(Boat boat, int row, int column)
+        {
+            if (boat.Orientation == Orientation.Horizontal && ValidHorizontalBoatLocation(boat, row, column))
+            {
+                Console.WriteLine("TRUE");
+                AddHorizontalBoat(boat, row, 0, column, 0);
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
+
+        private bool ValidHorizontalBoatLocation(Boat boat, int row, int column)
+        {
+            int boardRow = row - ROW_OFFSET + 1;
+            int boardColumn = TranslateConsoleGridToGameBoard(column);
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            {
+                if (GameBoard[boardRow, boardColumn + i] == BOAT_SPACE ||
+                    GameBoard[boardRow, boardColumn + i] == BOAT_SPACE_RIGHT_EDGE) 
+                { 
+                    return false;
+                };
+            }
+            return true;
+        }
+
+        public void AddHorizontalBoat(Boat boat, int row, int rowOffset, int column, int columnOffset)
+        { 
+            int boardRow = row - ROW_OFFSET + 1;
+            int boardColumn = TranslateConsoleGridToGameBoard(column);
+
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            { 
+                if (GameBoard[boardRow, boardColumn + i] == EMPTY_SPACE)
+                    GameBoard[boardRow, boardColumn + i] = BOAT_SPACE;
+                else
+                    GameBoard[boardRow, boardColumn + i] = BOAT_SPACE_RIGHT_EDGE;
+            }
+        }
+
+    private bool Collision(int row, int column)
+        {
+            return GameBoard[row, column] == BOAT_SPACE || GameBoard[row, column] == BOAT_SPACE_RIGHT_EDGE ? true : false;
+        }
+
+        public void ChangeOrientationAndShowBoatOnBoard(Boat boat, int row, int rowOffset, int column, int columnOffset)
+        {
+            if (boat.Orientation == Orientation.Horizontal)
+                EraseHorizontalBoat(boat, row, rowOffset, column, columnOffset);
+            else
+                EraseVerticalBoat(boat, row, rowOffset, column, columnOffset);
+            boat.ChangeOrientation();
+            ShowBoatOnBoard(boat, row, rowOffset, column, columnOffset);
+        }
+
+        public void ShowBoatOnBoard(Boat boat, int row, int rowOffset, int column, int columnOffset)
+        {
+            if (boat.Orientation == Orientation.Horizontal)
+                ShowHorizontalBoat(boat, row, rowOffset, column, columnOffset);
+            else
+                ShowVerticalBoat(boat, row, rowOffset, column, columnOffset);
+        }
+
+        public void ShowHorizontalBoat(Boat boat, int row, int rowOffset, int column, int columnOffset)
+        {
+            EraseHorizontalBoat(boat, row, rowOffset, column, columnOffset);
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            {
+                Display.WriteAt(Display.WriteUnderline("O"), row, column + i * 2);
+            }
+        }
+
+        public void EraseHorizontalBoat(Boat boat, int row, int rowOffset, int column, int columnOffset)
+        {
+            int boardRow = row - ROW_OFFSET + 1;
+            int boardColumn = TranslateConsoleGridToGameBoard(column);
+            int boardOffset;
+            if (columnOffset == 2) boardOffset = 1;
+            else if (columnOffset == -2) boardOffset = -1;
+            else boardOffset = 0;
+
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            {
+Display.WriteAt("withBoardOffset: " + (boardColumn + i - boardOffset), 15 + i, 0);
+                if (GameBoard[boardRow - rowOffset, boardColumn + i - boardOffset] == EMPTY_SPACE ||
+                    GameBoard[boardRow - rowOffset, boardColumn + i - boardOffset] == EMPTY_RIGHT_EDGE_SPACE)
+                    Display.WriteAt(Display.WriteUnderline(" "), row - rowOffset, column - columnOffset + (i * 2));
+                if (GameBoard[boardRow - rowOffset, boardColumn + i - boardOffset] == BOAT_SPACE ||
+                    GameBoard[boardRow - rowOffset, boardColumn + i - boardOffset] == BOAT_SPACE_RIGHT_EDGE)
+                    Display.WriteAt(Display.WriteUnderline("O"), row - rowOffset, column - columnOffset + (i * 2));
+            }
+        }
+
+        private void ShowVerticalBoat(Boat boat, int row, int rowOffset, int column, int columnOffset)
+        {
+            EraseVerticalBoat(boat, row, rowOffset, column, columnOffset);
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            {
+                Display.WriteAt(Display.WriteUnderline("O"), row + i, column);
+            }
+        }
+
+        public void EraseVerticalBoat(Boat boat, int row, int rowOffset, int column, int columnOffset)
+        {
+            int boardRow = row - ROW_OFFSET + 1;
+            int boardColumn = TranslateConsoleGridToGameBoard(column);
+            int boardOffset;
+            if (rowOffset == 1) boardOffset = 1;
+            else if (rowOffset == -1) boardOffset = -1;
+            else boardOffset = 0;
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            {
+                if (GameBoard[boardRow + i - boardOffset, boardColumn] == EMPTY_SPACE ||
+                    GameBoard[boardRow + i - boardOffset, boardColumn] == EMPTY_RIGHT_EDGE_SPACE)
+                    Display.WriteAt(Display.WriteUnderline(" "), row - rowOffset + i, column - columnOffset);
+
+                else if (GameBoard[boardRow + i - boardOffset, boardColumn] == BOAT_SPACE ||
+                    GameBoard[boardRow + i - boardOffset, boardColumn] == BOAT_SPACE_RIGHT_EDGE)
+                    Display.WriteAt(Display.WriteUnderline("O"), row - rowOffset + i, column - columnOffset);
+            }
+        }
+
+        private int TranslateConsoleGridToGameBoard(int column)
+        {
+            column = column - COLUMN_OFFSET;
+            if (column == 0) return column + 1;
+
+            int translator = 0;
+            for (int i = 2; i < column; i += 2) 
+            {
+                translator++;
+            }
+            return column - translator;
+        }
     }
 }