+++ PlaceBoatBehaviors/ComputerBoatPlacer.cs
@@ -6,7 +6,7 @@ namespace battleship.PlaceBoatBehaviors
 {
     class ComputerBoatPlacer : IPlaceBoatBehavior
     {
-        public void PlaceBoat()
+        public void PlaceBoat(Board board)
         {
             Console.WriteLine("Computer placing a boat");
         }