+++ PlaceBoatBehaviors/HumanBoatPlacer.cs
@@ -6,58 +6,83 @@ namespace battleship.PlaceBoatBehaviors
 {
     public class HumanBoatPlacer : IPlaceBoatBehavior
     {
-        public void PlaceBoat()
+        public void PlaceBoat(Board board)
         {
-            int boatTypes = Enum.GetValues(typeof(BoatType)).Length;
-            Console.WriteLine(boatTypes);
+            int numberOfBoats = Enum.GetValues(typeof(BoatType)).Length;
             ConsoleKeyInfo userInput;
             bool selectionWasMade;
             Boat boat;
-            int row;
-            int column;
+            int row, column;
 
-            for (int i = 0; i < boatTypes; i++)
+            for (int i = 0; i < numberOfBoats; i++)
             {
                 selectionWasMade = false;
                 boat = SimpleBoatFactory.INSTANCE.CreateBoat((BoatType)i);
                 row = Board.ROW_OFFSET;
                 column = Board.COLUMN_OFFSET;
-                Display.WriteAt(Display.WriteUnderline("X"), row, column);
+                board.ShowBoatOnBoard(boat, row, 0, column, 0);
 
                 while (!selectionWasMade)
                 {
                     userInput = Console.ReadKey(true);
                     if (userInput.Key == ConsoleKey.Enter)
                     {
-                        selectionWasMade = true;
+                        selectionWasMade = board.AddBoatToBoard(boat, row, column);
                     }
                     else if (userInput.Key == ConsoleKey.LeftArrow && column > Board.COLUMN_OFFSET)
                     {
-                        Display.WriteAt(Display.WriteUnderline(" "), row, column);
-                        column -= 2;
+                        board.ShowBoatOnBoard(boat, row, 0, column -= 2, -2);
                     }
                     else if (userInput.Key == ConsoleKey.UpArrow && row > Board.ROW_OFFSET)
                     {
-                        Display.WriteAt(Display.WriteUnderline(" "), row, column);
-                        row--;
+                        board.ShowBoatOnBoard(boat, --row, -1, column, 0);
                     }
-                    else if (userInput.Key == ConsoleKey.RightArrow && column < Board.COLUMN_OFFSET + 18)
+                    else if (userInput.Key == ConsoleKey.RightArrow && IsInRightBoundary(boat, column))
                     {
-                        Display.WriteAt(Display.WriteUnderline(" "), row, column);
-                        column += 2;
+                        board.ShowBoatOnBoard(boat, row, 0, column += 2, 2);
                     }
-                    else if (userInput.Key == ConsoleKey.DownArrow && row < Board.ROW_OFFSET + 9)
+                    else if (userInput.Key == ConsoleKey.DownArrow && IsInBottomBoundary(boat, row))
                     {
-                        Display.WriteAt(Display.WriteUnderline(" "), row, column);
-                        row++;
+                        board.ShowBoatOnBoard(boat, ++row, 1, column, 0);
                     }
                     else if (userInput.Key == ConsoleKey.Spacebar)
                     {
-                        boat.ChangeOrientation();
+                        if ((boat.Orientation == Orientation.Horizontal && WillBeInBottomBoundary(boat, row)) ||
+                            (boat.Orientation == Orientation.Vertical && WillBeInRightBoundary(boat, column)))
+                        {
+                            board.ChangeOrientationAndShowBoatOnBoard(boat, row, 0, column, 0);
+                        }
                     }
-                    Display.WriteAt(Display.WriteUnderline("X"), row, column);
                 }
             }
+        }
+
+        private static bool IsInRightBoundary(Boat boat, int column)
+        {
+            int HORIZONTAL_BOAT_RIGHT_BOUNDARY = Board.COLUMN_OFFSET + 20 - boat.BoatLives.Length * 2;
+            int VERTICAL_BOAT_RIGHT_BOUNDARY = Board.COLUMN_OFFSET + 18;
+            return boat.Orientation == Orientation.Horizontal && column < HORIZONTAL_BOAT_RIGHT_BOUNDARY ||
+                        boat.Orientation == Orientation.Vertical && column < VERTICAL_BOAT_RIGHT_BOUNDARY;
+        }
+
+        private static bool IsInBottomBoundary(Boat boat, int row)
+        {
+            int HORIZONTAL_BOAT_BOTTOM_BOUNDARY = Board.ROW_OFFSET + 9;
+            int VERTICAL_BOAT_BOTTOM_BOUNDARY = Board.ROW_OFFSET + 10 - boat.BoatLives.Length;
+            return boat.Orientation == Orientation.Horizontal && row < HORIZONTAL_BOAT_BOTTOM_BOUNDARY ||
+                       boat.Orientation == Orientation.Vertical && row < VERTICAL_BOAT_BOTTOM_BOUNDARY;
+        }
+
+        private static bool WillBeInRightBoundary(Boat boat, int column)
+        {
+            int HORIZONTAL_BOAT_RIGHT_BOUNDARY = Board.COLUMN_OFFSET + 21 - boat.BoatLives.Length * 2;
+            return column < HORIZONTAL_BOAT_RIGHT_BOUNDARY;
+        }
+
+        private static bool WillBeInBottomBoundary(Boat boat, int row)
+        {
+            int VERTICAL_BOAT_BOTTOM_BOUNDARY = Board.ROW_OFFSET + 11 - boat.BoatLives.Length;
+            return row < VERTICAL_BOAT_BOTTOM_BOUNDARY;
         }
     }
 }